+++ src/board_controller/openbci/ganglion_bglib/helpers.cpp
@@ -123,8 +123,8 @@ namespace GanglionLib
                 break;
             }
             auto end_time = std::chrono::high_resolution_clock::now ();
-            run_time =
-                std::chrono::duration_cast<std::chrono::seconds> (end_time - start_time).count ();
+            run_time = (int)std::chrono::duration_cast<std::chrono::seconds> (end_time - start_time)
+                           .count ();
         }
         return exit_code;
     }
