+++ src/board_controller/openbci/inc/ganglion.h
@@ -19,8 +19,8 @@ private:
 
     bool is_valid;
 
-    float const accel_scale = 0.016f;
-    float const eeg_scale = (1.2f * 1000000) / (8388607.0f * 1.5f * 51.0f);
+    double const accel_scale = 0.016f;
+    double const eeg_scale = (1.2f * 1000000) / (8388607.0f * 1.5f * 51.0f);
 
     std::string start_command;
     std::string stop_command;
